@@ -1,10 +1,7 @@
 package com.example.swingmusic;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.app.ActivityOptions;
 import android.content.Intent;
-import android.os.Build;
 import android.os.Bundle;
 import android.util.Pair;
 import android.view.View;
@@ -12,10 +9,10 @@ import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.firebase.FirebaseApp;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -28,28 +25,22 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
+
         getSupportActionBar().hide();
         this.getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
 
-        sign_in = (Button) findViewById(R.id.button1);
-        sign_up = (Button) findViewById(R.id.button2);
+        sign_in = findViewById(R.id.button1);
+        sign_up = findViewById(R.id.button2);
         img = findViewById(R.id.imageView);
         tvSwing = findViewById(R.id.textView);
         tvTag = findViewById(R.id.tagline);
 
-        sign_in.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                openActivitySignIn();
-            }
-        });
-        sign_up.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                openActivitySignUp();
-            }
-        });
+        sign_in.setOnClickListener(v -> openActivitySignIn());
+        sign_up.setOnClickListener(v -> openActivitySignUp());
+
+
+
     }
     public void closeActivity(View view){
         Intent intent = new Intent(MainActivity.this, SwingHomePage.class);
