@@ -10,9 +10,8 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.Window;
-import android.view.WindowManager;
 import android.widget.LinearLayout;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.android.material.appbar.MaterialToolbar;
@@ -21,10 +20,17 @@ import com.google.android.material.navigation.NavigationBarView;
 import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 
 public class SwingHomePage extends AppCompatActivity {
 
     BottomNavigationView bottomNavigationView;
+
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -38,6 +44,114 @@ public class SwingHomePage extends AppCompatActivity {
         MaterialToolbar toolbar = findViewById(R.id.toolbar);
         DrawerLayout drawerLayout = findViewById(R.id.drawer_lay);
         NavigationView navigationView = findViewById(R.id.ver_nav_view); // This is one is the side bar!
+
+        View headerView = navigationView.getHeaderView(0);
+
+        TextView nameTextView = headerView.findViewById(R.id.Acname);
+        TextView usernameTextView = headerView.findViewById(R.id.username);
+
+        FirebaseUser currentUser = FirebaseAuth.getInstance().getCurrentUser();
+        DatabaseReference userRef = FirebaseDatabase.getInstance().getReference().child("Users");
+
+        if(currentUser != null) {
+            String currentUserId = currentUser.getUid();
+            userRef.child(currentUserId).addListenerForSingleValueEvent(new ValueEventListener() {
+                @Override
+                public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                    if (dataSnapshot.exists()) {
+                        // Retrieve name and username from the dataSnapshot
+                        String name = dataSnapshot.child("name").getValue(String.class);
+                        String username = dataSnapshot.child("username").getValue(String.class);
+
+                        // Set the name and username to TextViews
+                        nameTextView.setText(name);
+                        usernameTextView.setText(username);
+                    }
+                }
+
+                public void onCancelled(@NonNull DatabaseError databaseError) {
+                    // Handle error
+                }
+            });
+        } else {
+            nameTextView.setText("New User");
+            usernameTextView.setText("Register");
+        }
+
+        MenuItem logoutMenuItem = navigationView.getMenu().findItem(R.id.logout);
+
+        logoutMenuItem.setOnMenuItemClickListener(new MenuItem.OnMenuItemClickListener() {
+            @Override
+            public boolean onMenuItemClick(@NonNull MenuItem menuItem) {
+                FirebaseAuth.getInstance().signOut();
+
+                Intent intent = new Intent(SwingHomePage.this,MainActivity.class);
+                startActivity(intent);
+                return true;
+            }
+        });
+
+
+        MenuItem rateUsMenuItem = navigationView.getMenu().findItem(R.id.nav_rateus);
+
+        rateUsMenuItem.setOnMenuItemClickListener(new MenuItem.OnMenuItemClickListener() {
+            @Override
+            public boolean onMenuItemClick(@NonNull MenuItem menuItem) {
+                String[] recipient = {"2020bcs007@sggs.ac.in"};
+                String subject = "User Rating of your SwingMusic app";
+                String message = "Give us the rating between 1 to 5\n\n\n" +
+                        "Thanks for your Rating!\n" +
+                        "Team SwingMusic";
+                Intent emailIntent = new Intent(Intent.ACTION_SEND);
+                emailIntent.setType("text/plain");
+                emailIntent.putExtra(Intent.EXTRA_EMAIL,recipient);
+                emailIntent.putExtra(Intent.EXTRA_SUBJECT,subject);
+                emailIntent.putExtra(Intent.EXTRA_TEXT,message);
+
+                if (emailIntent.resolveActivity(getPackageManager()) != null) {
+                    // Start the activity (email app)
+                    startActivity(emailIntent);
+                } else {
+                    // Handle the case where no email app is available
+                    Toast.makeText(SwingHomePage.this, "No email app found", Toast.LENGTH_SHORT).show();
+                }
+
+                return true;
+            }
+        });
+
+        MenuItem feedbackMenuItem = navigationView.getMenu().findItem(R.id.nav_feedback);
+        feedbackMenuItem.setOnMenuItemClickListener(new MenuItem.OnMenuItemClickListener() {
+            @Override
+            public boolean onMenuItemClick(@NonNull MenuItem menuItem) {
+                String[] recipient = {"2020bcs007@sggs.ac.in"};
+                String subject = "User Feedback of your SwingMusic app";
+                String message =  "Thanks for your time, we are glad to hear from you!\n\n" +
+                        "Please share your feedback below:\n\n" +
+                        "<!--Your feedback here--!>\n\n" +
+                        "Thanks for your feedback!\n" +
+                        "We are working hard to serve you!";
+
+                Intent emailIntent = new Intent(Intent.ACTION_SEND);
+                emailIntent.setType("text/plain");
+                emailIntent.putExtra(Intent.EXTRA_EMAIL,recipient);
+                emailIntent.putExtra(Intent.EXTRA_SUBJECT,subject);
+                emailIntent.putExtra(Intent.EXTRA_TEXT,message);
+
+                if (emailIntent.resolveActivity(getPackageManager()) != null) {
+                    // Start the activity (email app)
+                    startActivity(emailIntent);
+                } else {
+                    // Handle the case where no email app is available
+                    Toast.makeText(SwingHomePage.this, "No email app found", Toast.LENGTH_SHORT).show();
+                }
+
+                return true;
+            }
+        });
+
+
+
         toolbar.setNavigationOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -62,24 +176,22 @@ public class SwingHomePage extends AppCompatActivity {
                         startActivity(intent1);
                         break;
                     case R.id.nav_settings:
-                        Toast.makeText(SwingHomePage.this, "Settings is Clicked!", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(SwingHomePage.this, "settings is under work!", Toast.LENGTH_SHORT).show();
                         break;
+
                     case R.id.nav_aboutus:
-                        Intent intent2 = new Intent(SwingHomePage.this, MainActivity.class);
+                        Intent intent2 = new Intent(SwingHomePage.this,SwingAboutUsPage.class);
                         startActivity(intent2);
                         break;
-                    case R.id.logout:
-                        Toast.makeText(SwingHomePage.this, "Log Out is Clicked!", Toast.LENGTH_SHORT).show();
-                        break;
                     case R.id.nav_share:
-                        Toast.makeText(SwingHomePage.this, "Share is Clicked!", Toast.LENGTH_SHORT).show();
+                        String githubProfileUrl = "https://github.com/abhishek3345/";
+                        String shareMessage = "Hey! check out this SwingMusic app for listening songs. Download from this link: " + githubProfileUrl;
+                        Intent shareIntent = new Intent(Intent.ACTION_SEND);
+                        shareIntent.setType("text/plain");
+                        shareIntent.putExtra(Intent.EXTRA_TEXT, shareMessage);
+                        startActivity(Intent.createChooser(shareIntent, "Share via"));
                         break;
-                    case R.id.nav_rateus:
-                        Toast.makeText(SwingHomePage.this, "Rate Us is Clicked!", Toast.LENGTH_SHORT).show();
-                        break;
-                    case R.id.nav_feedback:
-                        Toast.makeText(SwingHomePage.this, "Feedback is Clicked!", Toast.LENGTH_SHORT).show();
-                        break;
+
                 }
                 return true;
             }
